refactor: consolidate finding renderer helpers
